requirements.txt
Django>=5.0
django-filter>=24.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
pandas>=2.2
numpy>=2.0
scikit-learn>=1.4
joblib>=1.3
reportlab>=4.0
xhtml2pdf>=0.2.16
openai>=1.0
google-generativeai>=0.8
python-dotenv>=1.0
